Variable types, Type checking, Keywords: final, const, dynamic
We can declare all variables with the generic type 'var'.
Variables declared without a static type are implicitly considered dynamic. So, they can also be Declared using the 'dynamic' keyword in place of the 'var' keyword.

Identifiers cannot contain spaces and special characters, except the underscore(_) and the dollar($) sign.
Variable names cannot begin with a number.

All variables in Dart store the REFERENCE to the variable rather than the value of the variable.

The final and const keywords are used to declare constants. Dart prevents modifying the values of a variable that's declared final or const.
Dart throws an exception if an attempt is made to modify variables declared final or const.