Syntax, Interpolation, String Properties, Manipulating strings,
String values in Dart can be represented using either single or double or triple quotes. Single line strings are represented using single or double quotes. Triple quotes are used to represent multi-line strings.

String Properties
The properties listed are all read-only.
1.	codeUnits
Returns an unmodifiable list of the UTF-16 code units of this string.
2.	isEmpty
Returns true if this string is empty.
3.	Length
Returns the length of the string including space, tab and newline characters.

Methods to Manipulate Strings
1.	toLowerCase()
Converts all characters in this string to lower case.
2.	toUpperCase()
Converts all characters in this string to upper case.
3.	trim()
Returns the string without any leading and trailing whitespace.
4.	compareTo()
Compares this object to another.
5.	replaceAll()
Replaces all substrings that match the specified pattern with a given value.
6.	split()
Splits the string at matches of the specified delimiter and returns a list of substrings.
7.	substring()
Returns the substring of this string that extends from startIndex, inclusive, to endIndex, exclusive.
8.	toString()
Returns a string representation of this object.
9.	codeUnitAt()
Returns the 16-bit UTF-16 code unit at the given index.