Inserting, Updating, Removing list items
Mutable Lists can grow dynamically at runtime. The List.add() function appends the specified value to the end of the List and returns a modified List object.
The List.addAll() function accepts multiple values separated by a comma and appends these to the List.

Dart also supports adding elements at specific positions in the List. The insert() function accepts a value and inserts it at the specified index. 
Similarly, the insertAll() function inserts the given list of values, beginning from the index specified.

Dart allows modifying the value of an item in a List. In other words, one can re-write the value of list item.
The List class from the dart:core library provides the replaceRange() function to modify List items. This function replaces the value of the elements within the specified range.

The List.remove() function removes the first occurrence of the specified item in the list. This function returns true if the specified value is removed from the list.
The List.removeAt function removes the value at the specified index and returns it.
The List.removeLast() function pops and returns the last item in the List. 
The List.removeRange() function removes the items within the specified range.