Map literals, Map constructors, Map properties and functions
The Map object is a simple key/value pair. Keys and values in a map may be of any type. A Map is a dynamic collection. In other words, Maps can grow and shrink at runtime.

Maps can be declared in two ways −
Using Map Literals
Using a Map constructor

A map value can be any object including NULL.

Map – Properties:
1.	Keys
Returns an iterable object representing keys
2.	Values
Returns an iterable object representing values
3.	Length
Returns the size of the Map
4.	isEmpty
Returns true if the Map is an empty Map
5.	isNotEmpty
Returns true if the Map is an empty Map.

Map - Functions:
1.	addAll()
Adds all key-value pairs of other to this map.
2.	clear()
Removes all pairs from the map.
3.	remove()
Removes key and its associated value, if present, from the map.
4.	forEach()
Applies f to each key-value pair of the map.