Runes
Since a Dart string is a sequence of UTF-16 code units, 32-bit Unicode values within a string are represented using a special syntax. A rune is an integer representing a Unicode code point.

String.codeUnitAt() Function
Code units in a string can be accessed through their indexes.

String.codeUnits Property
This property returns an unmodifiable list of the UTF-16 code units of the specified string.

String.runes Property
This property returns an iterable of Unicode code-points of this string.Runes extends iterable.