Declaring, Instantiating, Accessing data and functions
Instantiating:
The new keyword is responsible for instantiation.
The right-hand side of the expression invokes the constructor. The constructor should be passed values if it is parameterised.

Accessing Attributes and Functions:
A class’s attributes and functions can be accessed through the object. Use the ‘.’ dot notation (called as the period) to access the data members of a class.