Method overriding in dart
Method Overriding is a mechanism by which the child class redefines a method in its parent class.
The number and type of the function parameters must match while overriding the method. In case of a mismatch in the number of parameters or their data type, the Dart compiler throws an error. "Invalid override. The type of 'Child.m1' ('(String) → void') isn't a subtype of 'Parent.m1' ('(int) → void')."