Defining HashMaps, remove, clear
A HashMap is a hash table based implementation of Map. When you iterate through a HashMap's keys or values, you cannot expect a certain order.
The HashMap class inherits the addAll() function from the Map class. This function enables adding multiple values all at once.

The remove() and the clear() functions are used to remove entries from the HashMap. The remove() function is passed a key that represents the entry to be removed. The clear() function is used to remove all the entries from the Map.